+++ src/main/java/gs/mclo/api/internal/Initializable.java
@@ -1,0 +1,7 @@
+package gs.mclo.api.internal;
+
+import gs.mclo.api.MclogsClient;
+
+public interface Initializable {
+    void setClient(MclogsClient client);
+}